web-server/requirements1.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports-zoneinfo==0.2.1
cachetools==5.3.3
captcha==0.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.4
colorlog==4.7.2
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
django==3.2.5
django-cors-headers==4.3.1
django-filter==23.5
django-simpleui==2024.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
greenlet==3.0.3
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.4.0
inflection==0.5.1
intel-openmp==2021.4.0
itypes==1.2.0
jieba==0.42.1
jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
markdown==3.6
markupsafe==2.1.5
matplotlib==3.7.5
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mysqlclient==2.1.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
packaging==24.0
pandas==2.0.3
pillow==10.3.0
plotly==5.22.0
protobuf==5.27.1
psutil==5.9.8
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pyjwt==2.8.0
pymysql==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
recbole==1.1.1
redis==5.0.3
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
sqlalchemy==2.0.31
sqlparse==0.4.4
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.4.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1-2209072238
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torch-cluster==1.6.3+pt23cpu
torch-geometric==2.5.3
torch-scatter==2.1.2+pt23cpu
torch-sparse==0.6.18+pt23cpu
torch-spline-conv==1.2.2+pt23cpu
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typing-extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2